Cut over to the new GenericNumber values, deleting all the old stuff

+++ h2data/src/bitmasks/mod.rs
@@ -10,7 +10,7 @@ use lazy_static::lazy_static;
 ///
 /// The numeric column must be unique, but not every bit between 0 and 63 need
 /// be represented.
-fn load_from_csv(data: &str) -> SimpleResult<HashMap<u64, String>> {
+fn load_from_csv(data: &str) -> SimpleResult<HashMap<usize, String>> {
     let mut out = HashMap::new();
 
     let mut rdr = csv::ReaderBuilder::new()
@@ -26,7 +26,7 @@ fn load_from_csv(data: &str) -> SimpleResult<HashMap<u64, String>> {
             bail!("Bad bitmask CSV: must be 2 records per line, this line was {}", record.len());
         }
 
-        let number: u64 = record.get(0).ok_or(
+        let number: usize = record.get(0).ok_or(
             SimpleError::new("Error reading the CSV file")
         )?.parse().map_err(|e| {
             SimpleError::new(format!("Couldn't parse first CSV field as integer: {}", e))
@@ -46,7 +46,7 @@ fn load_from_csv(data: &str) -> SimpleResult<HashMap<u64, String>> {
 
 lazy_static! {
     /// Pre-load the BITMASKS structure
-    pub static ref BITMASKS: HashMap<String, HashMap<u64, String>> = {
+    pub static ref BITMASKS: HashMap<String, HashMap<usize, String>> = {
         let mut h = HashMap::new();
         h.insert("TerrariaVisibility".to_string(), load_from_csv(include_str!("./terraria_visibility.csv")).unwrap());
 
@@ -62,12 +62,12 @@ pub fn bitmask_exists(name: &str) -> bool {
 /// Convert from a 64-bit value to a list of booleans and bits.
 ///
 /// The return type is a list of tuples. The tuples are composed of three
-/// values: `(bool is_set, u64 bit_mask, String name)`.
+/// values: `(bool is_set, usize bit_mask, String name)`.
 ///
 /// * `is_set` - true if the value is on, false if it is off
 /// * `bit_mask` - the shifted integer value - bit `3` would be `(1<<3)` or `0x08`, for example
 /// * `name` - the name of the field, directly from the original CSV file
-pub fn from_bitmask(bitmask: &str, value: u64) -> SimpleResult<Vec<(bool, u64, String)>> {
+pub fn from_bitmask(bitmask: &str, value: usize) -> SimpleResult<Vec<(bool, usize, String)>> {
     let bitmask = BITMASKS.get(bitmask).ok_or(
         SimpleError::new(format!("No such bitmask: {}", bitmask))
     )?;
@@ -92,7 +92,7 @@ pub fn from_bitmask(bitmask: &str, value: u64) -> SimpleResult<Vec<(bool, u64, S
 /// Convert from a 64-bit value to a list of strings.
 ///
 /// "Negative" strings are only included if `include_negatives` is set.
-pub fn from_bitmask_str(bitmask: &str, value: u64, include_negatives: bool) -> SimpleResult<Vec<String>> {
+pub fn from_bitmask_str(bitmask: &str, value: usize, include_negatives: bool) -> SimpleResult<Vec<String>> {
     let mut out: Vec<_> = from_bitmask(bitmask, value)?.iter()
         .filter(|(is_set, _, _)| include_negatives || *is_set)
         .map(|(is_set, _, name)| {
